requirements.txt
alembic==1.12.1
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.0.12
click==8.1.7
cryptography==43.0.3
enum34==1.1.10
Flask==2.0.1
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Migrate==4.0.7
Flask-SQLAlchemy==2.5.1
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
Mako==1.3.6
MarkupSafe==2.1.5
psycopg2==2.9.10
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.1.0
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.47
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==2.2.0
zipp==3.20.2
